Refactoring en solution restructuring
