+++ README.md
@@ -1,19 +1,13 @@
 # SampleDXPClients
-Sample Clients to interact with the FANC Data eXchange Platform.
+Sample Clients to interact with the FANC Data eXchange Platform API.
 **Remark:** This code is only usable by approved FANC partners who obtained credentials to connect to the FANC DXP API.
 
-## API v1
+### SamplePHIClient
 
-### SampleControlBodyClient
-
-Contains sample code for FANC approved control bodies to interact with the FANC DXP API
-
-- SampleControlBodyClient -> New developments should be based on this sample code.
-- ~~SampleControlBodyLegacyClient -> Uses the legacy API. New developments should avoid using this.~~ Legacy client support has ceased.
-
-## API v2
+Contains sample code for FANC approved partners of the Physical Inventory Project (PHI)
 
 ### SampleJSClient
 
-Contains sample code on how to interact with the FANC API v2 via the means of a JavaScript client.
+Contains sample code on how to interact with the new FANC API v2 via the means of a JavaScript client.
 **Important:** make sure you run the project locally from url https://localhost:5001
+**The v2 API will replace the v1 API in the future. For internal usage only for the time being**
