+++ SampleDXPClients.sln
@@ -1,21 +1,17 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27130.2027
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29503.13
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SampleControlBodyClient", "src\API_v1\SampleControlBodyClient\SampleControlBodyClient.csproj", "{935DE2FA-2F8B-4512-919B-569C70AA25EF}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "API_v2", "API_v2", "{8C92FF71-CD93-415C-B142-B9D4925AB60F}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "API_v1", "API_v1", "{A1AE5A15-309D-4DBA-A1D2-109B145312B6}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SampleJSClient", "src\API_v2\SampleJSClient\SampleJSClient.csproj", "{0AD306B1-857C-4865-906F-25F91AE24999}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{DB3219B2-939C-4A01-9B63-1E4A0FDB83B0}"
 	ProjectSection(SolutionItems) = preProject
 		LICENSE = LICENSE
 		README.md = README.md
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SamplePHIClient", "src\SamplePHIClient\SamplePHIClient.csproj", "{935DE2FA-2F8B-4512-919B-569C70AA25EF}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SampleJSClient", "src\SampleJSClient\SampleJSClient.csproj", "{3A12ACC1-545E-4F56-A3DD-8D0387242791}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,17 +23,13 @@ Global
 		{935DE2FA-2F8B-4512-919B-569C70AA25EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{935DE2FA-2F8B-4512-919B-569C70AA25EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{935DE2FA-2F8B-4512-919B-569C70AA25EF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0AD306B1-857C-4865-906F-25F91AE24999}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0AD306B1-857C-4865-906F-25F91AE24999}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0AD306B1-857C-4865-906F-25F91AE24999}.Release|Any CPU.ActiveCfg = Debug|Any CPU
-		{0AD306B1-857C-4865-906F-25F91AE24999}.Release|Any CPU.Build.0 = Debug|Any CPU
+		{3A12ACC1-545E-4F56-A3DD-8D0387242791}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3A12ACC1-545E-4F56-A3DD-8D0387242791}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3A12ACC1-545E-4F56-A3DD-8D0387242791}.Release|Any CPU.ActiveCfg = Debug|Any CPU
+		{3A12ACC1-545E-4F56-A3DD-8D0387242791}.Release|Any CPU.Build.0 = Debug|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{935DE2FA-2F8B-4512-919B-569C70AA25EF} = {A1AE5A15-309D-4DBA-A1D2-109B145312B6}
-		{0AD306B1-857C-4865-906F-25F91AE24999} = {8C92FF71-CD93-415C-B142-B9D4925AB60F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7DE15089-10B3-4269-B189-5C0255E9668F}
