+++ src/SampleJSClient/SampleJSClient.csproj
