+++ src/SampleJSClient/libman.json
