+++ src/SampleJSClient/wwwroot/app.js
