+++ src/SampleJSClient/wwwroot/index.html
