+++ src/SampleJSClient/wwwroot/lib/oidc-client/oidc-client.js
