+++ src/SamplePHIClient/Properties/AssemblyInfo.cs
@@ -5,12 +5,12 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("SampleControlBodyClient")]
+[assembly: AssemblyTitle("SamplePHIClient")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("FANC-AFCN")]
-[assembly: AssemblyProduct("SampleControlBodyClient")]
-[assembly: AssemblyCopyright("Copyright © FANC-AFCN 2017")]
+[assembly: AssemblyProduct("SamplePHIClient")]
+[assembly: AssemblyCopyright("Copyright © FANC-AFCN 2019")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
