+++ src/SamplePHIClient/Properties/Resources.resx
