+++ src/SamplePHIClient/Properties/Settings.settings
